+++ untitled-sdl-tank-game/Player.cpp
@@ -1,6 +1,6 @@
 #include "Player.h"
 
-Player :: Player(int x, int y) : GameObject(x, y) {
+Player :: Player(float x, float y) : GameObject(x, y) {
 	posX = x;
 	posY = y;
 	myTex = NULL;
